Remove container in killStoreNodes
In TR::LocalDeadStoreElimination::killStoreNodes, a TR::SparseBitVector
is created to hold use def alias information which is used to do the
check to see if a store node can be killed. However, this increases
compilation time compared to the original way of doing this check,
which was to check the symref manually. The reason for this is because
it takes time to create the bit vector, and then the code ends up
iterating though the list of store nodes anyway, resulting in no real
benefit derived from prepopulating the bit vector.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>