Remove use of self() from hpp files
Using self() at the TR level doesn't make sense because the definition
of self() is to cast the this pointer of an extensible object to
TR::<extensible class name> * . As such, any method of an object in the
TR namespace that calls self() does a pointless cast. In addition, when
trying to inline self(), this can cause compiler errors because
depending on the include order, the type in question may not be a
complete type.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>